Create Note_Book
Create a class Book representing a book in a library. A book has a title (String), an author (String), a field to store the year when it was published (int), a field for the number of pages (int), and a logical value describing whether the book can be borrowed.

Create a constructor which requires all the above attributes of a book. Create another constructor which takes a Book reference as parameter. (Do not forget about the case when the parameter is null.)

Create a method toString() to convert the book data into a String. Create a sameAs(Book) method which returns true if the parameter book has the same title, author, year and pages as the current book. (Whether the books can be borrowed does not matter when comparing them.)

Create a main program to illustrtate the use of the class. Put all classes into the package library.